interview test
You are required to build a picture upload application.
A user must be able to upload pictures of format jpg, gif or png
Once picture has been uploaded the system need to confirm upload with the picture type.
The picture must be added to the list of picture thumbnails resized to 150px by 150px.
When l click the picture l must see a larger picture of 1000px by 1009px.
User must be able to filter the pictures by type.
The application needs to have a back end of PHP or NodeJS.
The application needs to contain a front end of Angular or Ionic.